Providers are not avaiable in OSGi containers. Moved package jsonpath.internal.spi to jsonpath.spi.
