+++ json-path/src/main/java/com/jayway/jsonpath/Configuration.java
@@ -14,8 +14,8 @@
  */
 package com.jayway.jsonpath;
 
-import com.jayway.jsonpath.internal.spi.json.JsonSmartJsonProvider;
-import com.jayway.jsonpath.internal.spi.mapper.JsonSmartMappingProvider;
+import com.jayway.jsonpath.spi.json.JsonSmartJsonProvider;
+import com.jayway.jsonpath.spi.mapper.JsonSmartMappingProvider;
 import com.jayway.jsonpath.spi.json.JsonProvider;
 import com.jayway.jsonpath.spi.mapper.MappingProvider;
 import org.slf4j.Logger;
