+++ json-path/src/main/java/com/jayway/jsonpath/spi/json/GsonJsonProvider.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jayway.jsonpath.internal.spi.json;
+package com.jayway.jsonpath.spi.json;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
