+++ json-path/src/main/java/com/jayway/jsonpath/spi/mapper/Factory.java
@@ -1,4 +1,4 @@
-package com.jayway.jsonpath.internal.spi.mapper;
+package com.jayway.jsonpath.spi.mapper;
 
 public interface Factory<T> {
