+++ json-path/src/test/java/com/jayway/jsonpath/BaseTest.java
@@ -1,11 +1,11 @@
 package com.jayway.jsonpath;
 
 import com.jayway.jsonpath.internal.Path;
-import com.jayway.jsonpath.internal.spi.json.GsonJsonProvider;
-import com.jayway.jsonpath.internal.spi.json.JacksonJsonNodeJsonProvider;
-import com.jayway.jsonpath.internal.spi.json.JacksonJsonProvider;
-import com.jayway.jsonpath.internal.spi.mapper.GsonMappingProvider;
-import com.jayway.jsonpath.internal.spi.mapper.JacksonMappingProvider;
+import com.jayway.jsonpath.spi.json.GsonJsonProvider;
+import com.jayway.jsonpath.spi.json.JacksonJsonNodeJsonProvider;
+import com.jayway.jsonpath.spi.json.JacksonJsonProvider;
+import com.jayway.jsonpath.spi.mapper.GsonMappingProvider;
+import com.jayway.jsonpath.spi.mapper.JacksonMappingProvider;
 import com.jayway.jsonpath.internal.token.PredicateContextImpl;
 
 import java.util.HashMap;
